initial commit. WIP.  TODO: will out implementations for Windows. Test and debug linux implementation.
