+++ src/time_linux.h
@@ -1,12 +1,5 @@
 
-
-
-#include <stdio.h>
-#include <stdlib.h>
 #include <sys/time.h>       /* gettimeofday */
-#include <mach/mach_time.h> /* mach_absolute_time */
-#include <mach/mach.h>      /* host_get_clock_service, mach_... */
-#include <mach/clock.h>     /* clock_get_time */
 
 #define BILLION 1000000000L
 #define MILLION 1000000L
@@ -19,6 +12,7 @@
         ts.tv_nsec = ts.tv_nsec % BILLION;              \
     } while (0)
 
+
 /*
  * Get absolute future time for pthread timed calls
  *  Solution 1: microseconds granularity
@@ -26,9 +20,7 @@
 struct timespec get_abs_future_time_coarse(unsigned milli)
 {
     struct timespec future;         /* ns since 1 Jan 1970 to 1500 ms in the future */
-    struct timeval  micro = {0, 0}; /* 1 Jan 1970 */
-
-    (void) gettimeofday(&micro, NULL);
+    clock_gettime(CLOCK_REALTIME, &future);
     future.tv_sec = micro.tv_sec;
     future.tv_nsec = micro.tv_usec * 1000;
     NORMALISE_TIMESPEC( future, milli );
